reducers put in there own folder
